Allowed Nodes to be double and reflected those changes to heuristics

+++ src/main/java/com/jaiPatel/aisearch/heuristics/EuclideanHeuristic.java
@@ -5,8 +5,8 @@ import com.jaiPatel.aisearch.graph.Node;
 public class EuclideanHeuristic implements Heuristic {
     @Override
     public double estimate(Node current, Node goal) {
-        int dx = current.getX() - goal.getX();
-        int dy = current.getY() - goal.getY();
+        double dx = current.getX() - goal.getX();
+        double dy = current.getY() - goal.getY();
         return Math.sqrt(dx * dx + dy * dy);
     }
 }
